Divided controller into 4 classes; Refactored ClientController;
