+++ src/main/java/com/andersen_intensive/hotel/controller/UtilityController.java
@@ -1,0 +1,4 @@
+package com.andersen_intensive.hotel.controller;
+
+public class UtilityController {
+}
